updated war machine

+++ ft_bzero.c
@@ -6,19 +6,19 @@
 /*   By: ckurt <ckurt@student.42lyon.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 14:08:58 by ckurt             #+#    #+#             */
-/*   Updated: 2020/11/23 17:31:08 by ckurt            ###   ########lyon.fr   */
+/*   Updated: 2020/11/24 10:26:57 by ckurt            ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
+#include "libft.h"
 
 void				ft_bzero(void *s, size_t n)
 {
 	unsigned char	*str;
-	int				i;
+	unsigned long	i;
 
 	i = 0;
 	str = (unsigned char *)s;
-	while ((unsigned long)i < n)
+	while (i < n)
 		str[i++] = '\0';
 }
